requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.4
django-crispy-forms==1.14.0
django-on-heroku==1.1.2
gunicorn==20.1.0
Pillow==9.1.1
psycopg2-binary==2.9.3
sqlparse==0.4.2
whitenoise==6.1.0